feat: refactor OrderList component and update import paths for improved structure and clarity
